[close #622] Move import_path() method into mixin
- create ImportPathMixin with import_path() method
- remove import_path() from Study
- remove import_path() from System
